doing put and minor changes to cgi response

+++ bin/create_file.py
@@ -4,8 +4,6 @@ from pathlib import Path
 import sys
 import json
 
-print("READING NOW")
-
 p = Path("files")
 
 data = sys.stdin.read()
@@ -16,7 +14,16 @@ content = request["content"]
 
 file_path = p / filename
 
-with file_path.open("w") as file:
-    file.write(content)
+if file_path.exists():
+	print("HTTP/1.1 409 Conflic\r")
+	print("Content-Type: application/json\r")
+	print("\n\r")
+	print(json.dumps({"status": "error", "message": "File already exists"}))
 
-print("From " + __file__ + " File created successfully: " + filename)
+else:
+	with file_path.open("w") as file:
+		file.write(content)
+	print("HTTP/1.1 201 Created\r")
+	print("Content-Type: application/json\r")
+	print("\n\r")
+	print(json.dumps({"status": "ok", "message": "File created"}))
